multiplication tables
This code generates multiplication tables from a starting number to an ending number. It first checks if the input start and end numbers are integers. If they are, it checks if the start number is less than or equal to the end number. If it is, it iterates through each number from start to end and calls the multiplicationTable function with each number. If the start number is greater than the end number, it throws an error message. If the input numbers are not integers, it also throws an error message.

The multiplicationTable function takes a number as an argument and generates the multiplication table for that number from 1 to 10. The result is printed on the console in the format num x count = num*count.

After printing the multiplication table for each number, the code also prints a message indicating the end of the multiplication table session.